Refactor joystick
