+++ src/joystick/joystick.c
@@ -25,7 +25,7 @@
 
 #include "pgcompat.h"
 
-#include "doc/joystick_doc.h"
+#include "joystick_doc.h"
 
 static pgJoystickObject *joylist_head = NULL;
 static PyTypeObject pgJoystick_Type;
